+++ src/main/java/com/stackmob/sdk/request/StackMobRequestWithPayload.java
@@ -17,6 +17,7 @@
 package com.stackmob.sdk.request;
 
 import com.stackmob.sdk.api.StackMob;
+import com.stackmob.sdk.api.StackMobOptions;
 import com.stackmob.sdk.api.StackMobSession;
 import com.stackmob.sdk.callback.StackMobRawCallback;
 import com.stackmob.sdk.callback.StackMobRedirectedCallback;
@@ -33,57 +34,56 @@ public class StackMobRequestWithPayload extends StackMobRequest {
     public StackMobRequestWithPayload(ExecutorService executor,
                                       StackMobSession session,
                                       HttpVerbWithPayload verb,
-                                      List<Map.Entry<String, String>> headers,
+                                      StackMobOptions options,
                                       List<Map.Entry<String, String>> params,
                                       String body,
                                       String method,
                                       StackMobRawCallback cb,
                                       StackMobRedirectedCallback redirCb) {
-        this(executor, session, null, verb, headers, params, body, method, cb, redirCb);
+        this(executor, session, null, verb, options, params, body, method, cb, redirCb);
     }
     public StackMobRequestWithPayload(ExecutorService executor,
                                       StackMobSession session,
                                       StackMob.OAuthVersion oauthVersionOverride,
                                       HttpVerbWithPayload verb,
-                                      List<Map.Entry<String, String>> headers,
+                                      StackMobOptions options,
                                       List<Map.Entry<String, String>> params,
                                       String body,
                                       String method,
                                       StackMobRawCallback cb,
                                       StackMobRedirectedCallback redirCb) {
-        super(executor, session, oauthVersionOverride, verb, headers, params, method, cb, redirCb);
+        super(executor, session, oauthVersionOverride, verb, options, params, method, cb, redirCb);
         this.body = body;
     }
 
     public StackMobRequestWithPayload(ExecutorService executor,
                                       StackMobSession session,
                                       HttpVerbWithPayload verb,
-                                      List<Map.Entry<String, String>> headers,
+                                      StackMobOptions options,
                                       List<Map.Entry<String, String>> params,
                                       Object requestObject,
                                       String method,
                                       StackMobRawCallback cb,
                                       StackMobRedirectedCallback redirCb) {
-        super(executor, session, null, verb, headers, params, method, cb, redirCb);
-        this.requestObject = requestObject;
+        this(executor, session, null, verb, options, params, requestObject, method, cb, redirCb);
     }
 
     public StackMobRequestWithPayload(ExecutorService executor,
                                       StackMobSession session,
                                       StackMob.OAuthVersion oauthVersionOverride,
                                       HttpVerbWithPayload verb,
-                                      List<Map.Entry<String, String>> headers,
+                                      StackMobOptions options,
                                       List<Map.Entry<String, String>> params,
                                       Object requestObject,
                                       String method,
                                       StackMobRawCallback cb,
                                       StackMobRedirectedCallback redirCb) {
-        super(executor, session, oauthVersionOverride, verb, headers, params, method, cb, redirCb);
+        super(executor, session, oauthVersionOverride, verb, options, params, method, cb, redirCb);
         this.requestObject = requestObject;
     }
 
     public StackMobRequestWithPayload(ExecutorService executor, StackMobSession session, StackMob.OAuthVersion oauthVersionOverride, HttpVerbWithPayload verb, String method, StackMobRawCallback cb, StackMobRedirectedCallback redirCb) {
-        this(executor, session, oauthVersionOverride, verb, EmptyHeaders, EmptyParams, null, method, cb, redirCb);
+        this(executor, session, verb, StackMobOptions.none(), EmptyParams, null, method, cb, redirCb);
     }
 
     @Override protected String getRequestBody() {