+++ src/main/java/com/stackmob/sdk/request/StackMobRequestWithoutPayload.java
@@ -17,6 +17,7 @@
 package com.stackmob.sdk.request;
 
 import com.stackmob.sdk.api.StackMob;
+import com.stackmob.sdk.api.StackMobOptions;
 import com.stackmob.sdk.api.StackMobSession;
 import com.stackmob.sdk.callback.StackMobRawCallback;
 import com.stackmob.sdk.callback.StackMobRedirectedCallback;
@@ -30,24 +31,24 @@ public class StackMobRequestWithoutPayload extends StackMobRequest {
     public StackMobRequestWithoutPayload(ExecutorService executor,
                                          StackMobSession session,
                                          HttpVerbWithoutPayload verb,
-                                         List<Map.Entry<String, String>> headers,
+                                         StackMobOptions options,
                                          List<Map.Entry<String, String>>  params,
                                          String method,
                                          StackMobRawCallback cb,
                                          StackMobRedirectedCallback redirCb) {
-        this(executor, session, null, verb, headers, params, method, cb, redirCb);
+        super(executor, session, null, verb, options, params, method, cb, redirCb);
     }
 
     public StackMobRequestWithoutPayload(ExecutorService executor,
                                          StackMobSession session,
                                          StackMob.OAuthVersion oauthVersionOverride,
                                          HttpVerbWithoutPayload verb,
-                                         List<Map.Entry<String, String>> headers,
+                                         StackMobOptions options,
                                          List<Map.Entry<String, String>>  params,
                                          String method,
                                          StackMobRawCallback cb,
                                          StackMobRedirectedCallback redirCb) {
-        super(executor, session, oauthVersionOverride, verb, headers, params, method, cb, redirCb);
+        super(executor, session, oauthVersionOverride, verb, options, params, method, cb, redirCb);
     }
 
     public StackMobRequestWithoutPayload(ExecutorService executor,
@@ -57,7 +58,7 @@ public class StackMobRequestWithoutPayload extends StackMobRequest {
                                          String method,
                                          StackMobRawCallback cb,
                                          StackMobRedirectedCallback redirCb) {
-        this(executor, session, oauthVersionOverride, verb, EmptyHeaders, EmptyParams, method, cb, redirCb);
+        this(executor, session, oauthVersionOverride, verb, StackMobOptions.none(), EmptyParams, method, cb, redirCb);
     }
 
     @Override protected String getRequestBody() {